requirements.txt
asgiref==3.7.2
Django==4.2.3
django-countries==7.5.1
django-credit-cards==0.4.1
Pillow==10.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
